Move undoVersion to padding before crc, change to uint8
- Move undoVersion field before crc in CheckpointControl
- Change from uint32 to uint8 (only need values 0 and 1)
- Add explicit padding1 and padding2 bytes (always were there, now named)
- Place undoVersion at offset 157 (in old padding area)
- CRC remains at same offset (160) in both old and new structures
- Old files: undoVersion reads 0 from zeroed padding
- New files: undoVersion=1 explicitly set in padding

This preserves CRC offset and calculation range, ensuring perfect
backward compatibility. The padding bytes were always zero in old files.

+++ include/transam/undo.h
@@ -344,7 +344,7 @@ extern PendingTruncatesMeta *pending_truncates_meta;
 extern Size undo_shmem_needs(void);
 extern void undo_shmem_init(Pointer buf, bool found);
 extern UndoMeta *get_undo_meta_by_type(UndoLogType undoType);
-extern void set_cluster_undo_version(uint32 version);
+extern void set_cluster_undo_version(uint8 version);
 
 extern void update_min_undo_locations(UndoLogType undoType,
 									  bool have_lock,
